requirements.txt
bittensor==9.9.0
pandas
scikit-learn
wandb
tabulate
python-dotenv

# Additional dependencies
aiohttp
pydantic
numpy
joblib
